tools/requirements.txt
requests>=2.31
python-dotenv>=1.0
openai>=1.0
jinja2>=3.1.4
markdown>=3.5
fastembed>=0.8.0
pyarrow>=24.0.0
numpy>=2.0.0
Pillow>=10.0.0
